バックエンド機能を頑張りました

+++ dokoTsubu/src/main/java/model/DeleteMutterLogic.java
@@ -1,15 +1,17 @@
 package model;
 
-import dao.CommentsDAO;
 import dao.MuttersDAO;
 
 public class DeleteMutterLogic {
-    public void execute(int mutterId) {
-        CommentsDAO commentsDAO = new CommentsDAO();
-        commentsDAO.deleteByMutterId(mutterId); // 関連するコメントを削除
+    public boolean execute(int mutterId, String loginUserName) {
         MuttersDAO dao = new MuttersDAO();
-        dao.delete(mutterId);
+        String postUserName = dao.findUserByMutterId(mutterId);
+
+        if (postUserName == null || !postUserName.equals(loginUserName)) {
+            System.out.println("ユーザーが一致しません。削除できません。");
+            return false;
+        }
+
+        return dao.delete(mutterId);
     }
 }
-
-
